More moves towards using adapted STIX fonts
git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@3201 2bfe0521-f11c-4a00-b80e-6202646ff360

+++ csl/cslbase/wxfonts/adjustfont.sh
@@ -11,6 +11,16 @@
 # the standard way. If your system is set up differently you will need
 # to edit this!
 
+case "x$1"
+in
+*keep*)
+  keeptemp="yes"
+  ;;
+*)
+  keeptemp="no"
+  ;;
+esac
+
 if which fontforge > /dev/null 2> /dev/null
 then
   ff=fontforge
@@ -50,7 +60,10 @@ gcc fixafm.c -o fixafm
 for x in *.afm
 do
   ./fixafm $x metrics/$x
-  rm $x
+  if test "$keeptemp" = "no"
+  then
+    rm $x
+  fi
 done
 rm fixafm
